Create public API creator
`Api` is a struct which is required on several public interfaces on this
crate (for example, `Reflector` or to simply fetch objects of a specific
type).

By having this public interface, one would be able to create `Api`
structs for other types which are not distributed by this crate. For
example, if one wants to create an `Api` for a custom type, now is
impossible.

Another use case can be to wrap objects on a reference counted type (for
example `Rc<Object<PodSpec, PodStatus>>` or `Arc<Object<PodSpec,
PodStatus>>`). For example, using an `Arc` wrapped object, allows
extract objects from a `Reflector` on a cheaper way when objects are
large (as it avoids copy `Object`s that may be really large).