separate seralization type from the api resource type - #142
lets you use Api without the generic type parameter, but all the methods
have the generic type param.

this lets you re-use a `Api::namespaced::<Pod>(...)` (say) by doing:
- api.get::<Pod>(...)
- api.get::<LimitedPod>)(...)

which can be useful sometimes.

Not entirely sure about this, but it has a use-case for me where I need
to re-define a CRD type twice (one with the api-guaranteed fields, and
one without)