revamp ObjectRef interface + bump dashmap
objectref had issues due to multiple string args that could easily get
confused so made a tiny builder in the same style for it.

dashmap bump might not have been necessary (and it's a little
premature), but was worried version 4 was not going to work. had to add
'static lifetimes on the Store for it to work with the new version
though.