Store: Allow namespaced ObjectRefs to cluster-scoped objects
Fixes #293.

This is a bit magical, but it's fairly common in K8s to allow reference
fields (such as `OwnerReferences`) to refer to either cluster-scoped objects
or namespaced resources in the same namespace.

Code that converts those events into `ObjectRef`s generally won't know which of
these cases is applicable, so this is the best we can do.