Generate schema for CRD v1
Schema is generated with `schemars`.

`inline_subschemas` option is used to avoid definitions in schema.
`meta_schema` is set to `None` to prevent including `$schema` which is
not allowed by Kubernetes.

This support nested structs and enums, but does not support structural
schemas for some of the more complex spec types.
The schema of the status subresource is included if present.

Also fixed the location of `subresources` and `additionalPrinterColumns`.

Closes https://github.com/clux/kube-rs/issues/264