Update k8s-openapi
This change updates k8s-openapi to pick up Arnavion/k8s-openapi@0a9d7e4,
which implements a workarond for a bug in the Kubernetes Node API that
wherein empty lists are encoded as `null`. This workaround is necessary
until the Kubernetes v1.22 API is released including a change to the
OpenAPI spec in kubernetes/kubernetes#102159.

k8s-openapi has not yet released this fix to crates.io, so instead this
change uses a dependency patch to pull in the latest version from git.