doc = include_str! readme
minimal tests for the readme.
trying to follow: https://blog.guillaume-gomez.fr/articles/2021-08-03+Improvements+for+%23%5Bdoc%5D+attributes+in+Rust

it captures invalid imports root imports, but not full paths for some
reason. maybe it gets confused with the multiple examples? i tried to
put it in a #![doc] module comment instead, but no change

Signed-off-by: clux <sszynrae@gmail.com>