runtime: Store resources an `Arc`
The read-side of `kube_runtime::reflector`  is allocation-heavy: each
read incurs a clone of the entire resource data structure. This can
incur substantial unnecessary load in read-heavy use cases, especially
for larger resources. Furthermore, with f66dd90, these allocations are
frequently performed while a read-lock is held.

This change modifies the `kube_runtime::reflector` to store resources in
an atomic, reference-counted cell (`std::sync::Arc`) so that clones do
not incur allocation. The goal of this change is to amortize costs so
that these allocations are only incurred once--at update-time, before
the write-lock is acquired--making reads cheaper and generally reducing
lock contention.

This changes the public APIs of `kube_runtime`'s `controller`,
`reflector`, and `finalizer` modules.

Signed-off-by: Oliver Gould <ver@buoyant.io>