Add a VS Code devcontainer configuration
When loading `kube-rs` in VS Code, a variety of configurations are
needed to end up with a working environment. It does't Just Work out of
the box.

VS Code [devcontainers][dc] make it easy to bootstrap a development
environment with project-specific tooling and configuration. This
configuration should also be usable in [GitHub Codespaces][ghcs].

This configuration ensures that various tools are available including:

* `cargo-deny`
* `cargo-tarpaulin`
* `k3d`
* `kubectl`
* rust stable-1.58 toolchain
* rust nightly toolchain
* a variety of related VS Code plugins

While this developer image is based on rust 1.56 (the documented MSRV
for the project), it updates the default toolchain to be _nightly_ so
clippy and rustfmt work as expected (since `kube-rs` depends on nightly
versions of these tools).

[dc]: https://code.visualstudio.com/docs/remote/containers
[ghcs]: https://github.com/features/codespaces

Signed-off-by: Oliver Gould <ver@buoyant.io>