portforward: Propose API changes to support cancelation
The `portforward` utility spawns a background task and returns a future
that completes when the background task completes, but the spawned
task's `JoinHandle` is discarded, so there's no way to cancel the
background tasks before all of the port forwards complete.

Furthermore, the `Portforwarder` types holds all `Port` instances and
exposes no APIs to allow a caller to take ownership. This makes it
difficult to work with a port stream. For example, it can't be moved
onto a spawned task directly. Instead a user has to extract the streams
(and errors) from the data structure.

To try to simplify this API:

* `Portforwarder` is updated to hold its streams and error handles in
  HashMaps keyed by port number. It is updated to expose `take_stream`
  and `take_error` methods so that callers can access either the I/O
  stream or the error handle by-numer. This is proposed to improve
  redability to avoid the unnecessary intermediate `Port` type.
* `Portforwarder` is updated to hold a `tokio::task::JoinHandle` and it
  now exposes `abort` and `join` methods that interact with the
  background task.
* `Portforward` no longer implements `Future` directly. This is done to
  reduce boilerplate code for what is now a very simple `async fn`.

Signed-off-by: Oliver Gould <ver@buoyant.io>