runtime: Update `await_condition` to return a value
Frequently when using `kube::runtime::wait::await_condition`, the caller
is waiting to obtain some value from the resource. For example, an
operator may wait to obtain a pod container's exit code. Currently, the
`await_condition` API does not permit returning a value from a
`Condition`, so callers must issue an additional API call to obtain the
value, even though the application has already observed it.

This change modifies the `Condition` trait to have an extra, optional
type parameter, `T`, so that `Condition::matches_object` returns an
`Option<T>` instead of a `bool` and `await_condition` now returns a
`Result<T, _>`.

This change avoids making any larger changes:

* `Condition<K>` remains a viable alias for `Condition<K, ()>`
* `Condition<K>` is implemented for `Fn(Option<&K>) -> bool`
* The implementations of `And` and `Not` are only implemented for
  `Condition<K>` (and not for other result types).
* The implementation of `Or` works with any two `Condition<K, T>`

This helps to minimize vebosity for existing use cases while also
satisfying other cases where a value is obtained.

Another up-side of this change is that the `?` directive can be used to
short-circuit matches.

The `kube::runtime::wait::condition_fn` helper is provided to construct
a `Condition` from a function that returns an optional value. This
helper is provided to distinguish from the default implementation for
`Fn(Option<&K>) -> bool`.

Signed-off-by: Oliver Gould <ver@buoyant.io>