runtime: Return the object in `await_condition`
Frequently when using `kube::runtime::wait::await_condition`, the caller
is waiting to obtain some value from the resource. For example, an
operator may wait to obtain a pod container's exit code. Currently, the
`await_condition` API does not permit returning a value from a
`Condition`, so callers must issue an additional API call to obtain the
value, even though the application has already observed it.

This change modifies the signature of `await_condition` to return an
`Option<K>` representing the state of the object when the condition was
satisfied. This eliminates the need for callers to issue an additional
request to get the state of the resource after the condition is met.

Signed-off-by: Oliver Gould <ver@buoyant.io>