fix applier hangs which can happen with many watched objects and children
When the applier scheduler channel buffer gets full, it will block preventing any further progress by the applier. This can be triggered with many objects which have many children that are being watched as the buffer fills up and progress is no longer made and re-queues can get to a point where they are never processed once the buffer is full.

Signed-off-by: Moustafa Baiou <moustafab.ccit@gmail.com>