Add support for metadata API
This change adds support for the (undocumented) metadata API. Through
this change, clients may request only the metadata of objects, reducing
I/O and storage pressure. To make it possible, the change adds new
request impl in kube-core, and new metadata methods on the Api type.

Additionally, this change also adds a new 'PartialObjectMeta' type that
is supposed to mirror the Go counterpart. A wrapper list has also been
added. Both of these support conversions to ObjectMeta (or
ObjectList<ObjectMeta>).

Metadata-only requests are made possible through the REST API by
including an extended Accept header. The header will notify the API
Server a JSON response is expected, and it will encode an additional
'as=PartialObjectMetadata' k=v pair to receive only the metadata of the
object. This is more efficient when only the metadata needs to be parsed
(as opposed to the whole spec).

Note: Older servers (k8s v1.14 and below) will retrieve the object and
then convert the metadata

Signed-off-by: Matei David <matei@buoyant.io>