Make `VersionMatch` follow upstream + decouple enum from `watcher`
Effectively makes the `VersionMatch` an exact duplicate of the upstream,
and adds some builders here and there to compensate for the missing `Any`.

Have changed the watcher interface and made a new enum that maps onto `VersionMatch` instead,
because it does not make sense to run a watcher against a pinned resource version; it will always use 0.

(at least unless/until we build pagination into it).

We could keep the enum that wraps resource version, but this is awkward;
it would be the only one that bundles resourceVersion that way;

- watch api calls get it as a str (outside watchparams)
- get api calls (in a follow-up pr) will also not use the enum

so altogether we are just making it awkward for ourselves by trying to make it nice :(

Signed-off-by: clux <sszynrae@gmail.com>