Add `Controller::for_stream`
This adds a stream interface analogue for `Controller::new`,
that allows the user to create the root `watcher` and `reflector`,
outside, and pass in the reader + flattened stream.

We originally discussed this name as `from_stream`, but given that we
need to also pass along the `reader`, it is no longer a `From` equivalent.
So have instead named it along the lines of `Controller::for` and highlighted
the terminology. Note that `Controller::For` is the [controller-runtime equivalent](https://pkg.go.dev/sigs.k8s.io/controller-runtime/pkg/builder#Builder.For).

Also fixes up bad doc setups and ensures unstable features are highlighted,
even though they appear in the middle of a trait.

The owns-stream unstable flagged is moved to `unstable-runtime-stream-share`
which now also contains this new method.

Signed-off-by: clux <sszynrae@gmail.com>