Fix `WatchParams` bookmarks for `watch_metadata`
Intended to factor out `populate_qp` from `ListParams` and do the same in `WatchParams`,
but accidentally found a place where the duplication caused a split.

This brings the query param serialization in line for all variants of list,
and does the same for all variants of watch (where the inconsistency was wound).

Signed-off-by: clux <sszynrae@gmail.com>