Swap `validator` for `garde`
The validator crate used in the derive docs and in the crd example is no
longer actively maintained. It has been superseded by garde which offers
similar functionality.

This change replaces validator with garde in dev dependencies (and in
the crd_api example dependency). Although garde has the same derive
macro, its attribute differs from validator, and it additionally
requires fields to be explictly skipped from validation, and does not
offer a built-in way to validate required fields.

Changes have been made to the kube_derive crate's docs to both replace
existing mentions of validator, and inform on the caveats mentioned
above.

Signed-off-by: Matei David <matei.david.35@gmail.com>