Swap `validator` for `garde` (#1212)
* Swap `validator` for `garde`

The validator crate used in the derive docs and in the crd example is no
longer actively maintained. It has been superseded by garde which offers
similar functionality.

This change replaces validator with garde in dev dependencies (and in
the crd_api example dependency). Although garde has the same derive
macro, its attribute differs from validator, and it additionally
requires fields to be explictly skipped from validation, and does not
offer a built-in way to validate required fields.

Changes have been made to the kube_derive crate's docs to both replace
existing mentions of validator, and inform on the caveats mentioned
above.

Signed-off-by: Matei David <matei.david.35@gmail.com>

* Disable default features for garde

Signed-off-by: Matei David <matei.david.35@gmail.com>

* Remove dev-dep in kube & kube-derive

Signed-off-by: Matei David <matei.david.35@gmail.com>

* Change validate attr to schemars

Signed-off-by: Matei David <matei.david.35@gmail.com>

* Update kube-derive/src/lib.rs

Co-authored-by: Eirik A <sszynrae@gmail.com>
Signed-off-by: Matei David <matei.david.35@gmail.com>

* Remove garde from kube examples

Signed-off-by: Matei David <matei.david.35@gmail.com>

* Re-work headers and remove garde

Signed-off-by: Matei David <matei.david.35@gmail.com>

---------

Signed-off-by: Matei David <matei.david.35@gmail.com>
Co-authored-by: Eirik A <sszynrae@gmail.com>