Introduce `GetParams` support (#1214)
* Introduce `GetParams` support

As part of a larger change to support version matching strategies (to
reduce I/O pressure), ListParams were separated into List and Watch
params. To complete the feature set, this change adds support for
`GetParams`, mirroing the upstream client-go layout for the type.

In the context of this change, we added a GetParams struct that does
implicit version matching (semantics are quite easy, any = "0",
NotOlderThan = specific, non-zero version, and unspecified is most
recent). We use the struct in all request methods and in Api core
methods. `ListParams` with version matching semantics is used for
list_metadata, so the same was done here (extended GetParams to
get_metadata and all other asssociated methods) even if it was outside
of the scope of the issue.

As part of the change, we also had to change the examples and tests.
Since the version was (largely) unspecified, we use the default derived
implementation; for tests and examples, this change should be
non-functional (unspecified is the same as unset).

Fixes #1174

Signed-off-by: Matei David <matei.david.35@gmail.com>

* Correctly quote resourceVersion in code docs

Signed-off-by: Matei David <matei.david.35@gmail.com>

* Add unit tests

Signed-off-by: Matei David <matei.david.35@gmail.com>

---------

Signed-off-by: Matei David <matei.david.35@gmail.com>