add `Duration` to `kube-core`
## Motivation

Some objects contain strings which represent durations. These strings
are often parsed in the format implemented by the Go standard library's
`time.Duration` type. For example,
https://github.com/kubernetes-sigs/gateway-api/pull/1997 adds duration
fields to the Gateway API's HTTPRoute type, and explicitly specifies
that they are formatted according to the Go `time.Duration` format:
> ```go
> // Timeout values are formatted like 1h/1m/1s/1ms as parsed by Golang time.ParseDuration
> // and MUST BE >= 1ms or 0 to disable (no timeout).
> ```
> (ref: https://github.com/kubernetes-sigs/gateway-api/blob/659d478ec039271a95429b133a999d1593808e96/geps/gep-1742.md?plain=1#L336-L337)

In the Go apimachinery's `meta` package, there is a `v1.Duration` struct
for representing durations, whose JSON representation is a string in the
Go `time.ParseDuration`/`time.Duration` format:
https://pkg.go.dev/k8s.io/apimachinery/pkg/apis/meta/v1#Duration.

That type is used in the Go bindings for Kubernetes objects that include
durations in this format, such as in the Gateway API proposal I
mentioned above:
> ```go
>        // +kubebuilder:validation:Format=duration
>        Request *metav1.Duration `json:"request,omitempty"`
> ```
> (ref: https://github.com/kubernetes-sigs/gateway-api/blob/659d478ec039271a95429b133a999d1593808e96/geps/gep-1742.md?plain=1#L353-L354)

## Solution

This branch adds a new `Duration` type to `kube-core`, which is
essentially an equivalent to `metav1.Duration`. This is a wrapper around
a Rust `std::time::Duration` which can be serialized and deserialized
using the same format as the Go `metav1.Duration` type. In addition,
because Go `time.Duration`s are represented by a signed number,
`Duration`s can also be negative if the provided duration string is
negative, and implement the correct ordering properties when compared to
each other and to `std::time::Duration`s. A parsed `Duration` can be
converted into a `std::time::Duration` by discarding its sign.

Closes #1222

Signed-off-by: Eliza Weisman <eliza@buoyant.io>