Fixup watcher docs, and watchlist doc examples
Lots of missing references in doc builds because many of these are warnings.

Removed the `namespace_reflector` because it's a slowly changing thing to watch, and instead added a way to opt-in to `WATCHLIST=1` in pod and node watchers.

Was trying to do this on event_watcher also but apparently events on k3s does not support sendInitialEvents on 1.27 yet so left it out there.
While in the area i also updated event_watcher to use the new `Event` type to make the example better.

Signed-off-by: clux <sszynrae@gmail.com>