client: Add feature flag for BoringSSL TLS
## Motivation

Curerently, `kube-client` supports using either Rustls or OpenSSL as the
TLS backend. BoringSSL is an alterative TLS implementation, based on a
fork of OpenSSL and maintained by Google. Rust bindings for BoringSSL
are available in the `boring` crate. Some users of `kube-client` may
wish to use BoringSSL as the TLS implementation, rather than OpenSSL or
Rustls. This can potentially be implemented in downstream code, on top
of `kube-client`'s config module, but it seemed nicer to just provide it
in `kube-client`.

## Solution

This commit adds support for BoringSSL as a third TLS option. The
`boring` crate has a very similar API to the `openssl` crate, so the
BoringSSL client code looks quite similar to the OpenSSL client code.

I've also attempted to modify the CI configuration to also run tests
with BoringSSL. However, I wasn't able to verify that this works without
an actual CI run, so it's possible that the E2E test docker environment
can't actually build BoringSSL and we'll have to add additional build
deps there. I'm happy to follow up on that after a CI run has completed.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>