Add example for raw API Server requests
I noticed a couple of questions around querying the metrics API that a
kubelet exposes. A number of different use cases aside from stat queries
may require raw requests to the API Server (in the style of `kubectl get --raw`).

It does not make much sense to extend the Request API to cover this use
case (since an abstraction would not simplify anything). Users can,
however, make use of the `http::Request` type and use an instantiated
client to interface with the API Server. There's a lack of material
here, so I thought I'd include an example that other people may
reference.

The example builds the equivalent of `kubectl top nodes` (in a simpler,
and dumber way) using raw HTTP requests. The data may not be as exact as
what `kubectl top` offers, but it serves as an example on how data may
be queried.

Signed-off-by: Matei David <matei.david.35@gmail.com>