Switch from `jsonpath_lib` to `jsonpath-rust` (#1345)
`jsonpath_lib` is unmaintained, which is not ideal.  It is also forcing
everyone to enable a `preserve_order` feature in the `serde_json`
dependency.

`preserve_order` is not following the proper feature model, where
features are only adding functionality.  It is actually changing how
`serde_json` works, switching from `BTreeMap` to `IndexMap`.

Signed-off-by: Illia Bobyr <illia.bobyr@gmail.com>
Co-authored-by: Eirik A <sszynrae@gmail.com>