cargo-release process seems solid enough to automate tag creation/pushing
we used to do this manually because failures were common from cargo-release
while waiting for the crate to be present on crates.io. this no longer
seems to be the case (particularly with the 20s wait time enforced),
so telling it to publish,tag,push all by itself to lessen number of manual steps.

Signed-off-by: clux <sszynrae@gmail.com>