feat(runtime): support for owned shared streams (#1483)
We added support for shared streams in the controller via a reflector.
There are only two public interfaces available to thread a stream
through a controller, `for_shared_stream` and `for_shared_stream_with`.

For non-shared streams, the controller can compose with other types of
streams that will trigger reconciliation, e.g. streams of owned objects,
or custom streams that take mappers out-of-band. For parity, the
controller should expose public interfaces to allow the same for shared
streams.

This change adds four new interfaces:

* `owns_shared_stream`
* `owns_shared_stream_with`
* `watches_shared_stream`
* `watches_shared_stream_with`

The contract is the same as `for_shared_stream`. Accept a stream of
`Arc<K>` and turn it into a reconciliation trigger.

---------

Signed-off-by: Matei David <matei.david.35@gmail.com>