Prevent empty string object name requests from being sent to the apiserver (#1541)
Fix #1540, due to the URL-nature a get with an empty name is processed as a list call.
kube then fails to parse the result from the server.
Other HTTP METHODS, like "DELETE" are also affected and e.g. cause the deletion of ALL objects. See the issue for more info.

Signed-off-by: Marcel Märtens <marcel.cochem@googlemail.com>