Experiment with multi_reflectors
since the question has showed up a few times.

ultimately this is awkward because the hard generic use in ObjectRef<K>
and in Store<K> which ties them to a particular resource.

It feels like it shouldn't have to be this way because the Store::get
takes an ObjectRef, which feels dynamically typed, but actually isnt.

Tried out a couple of potentials and had to just resort to the dumb thing instead.

Maybe there's some other things we can do for dynamics. Maybe worth exploring in an issue?

Signed-off-by: clux <sszynrae@gmail.com>