feat: Add typed scale argument to derive macro (#1656)
* feat: Add typed scale argument to derive macro

This allows cutomizing the scale subresource by providing key-value
items instead of a raw JSON string. For backwards-compatibility, it
is still supported to provide a JSON string. However, all examples
and tests were converted to the new format.

Signed-off-by: Techassi <git@techassi.dev>

* refactor: Remove k8s_openapi dependency

Signed-off-by: Techassi <git@techassi.dev>

* chore: Adjust doc comment, fix clippy lint

Signed-off-by: Techassi <git@techassi.dev>

* chore: Fix clippy lint in kube-runtime

Signed-off-by: Techassi <git@techassi.dev>

* docs: Adjust doc comment

Signed-off-by: Techassi <git@techassi.dev>

* chore: Use serde derive feature to enable derive macro

Signed-off-by: Techassi <git@techassi.dev>

* chore: Use ignore instead of no_run

Signed-off-by: Techassi <git@techassi.dev>

* test: Add schema test, fix FromMeta implementation

Adding this test proved to be very valuable because the FromMeta
implemenetation had a few errors and resulted in different panic
messages coming from the derive macro.

I also added a small note to the #[kube(scale(...))] section stating
that the scale subresource can only be used when the status subresource
is used as well. I plan to further improve the validation in a future
pull request.

Signed-off-by: Techassi <git@techassi.dev>

---------

Signed-off-by: Techassi <git@techassi.dev>