feat(deps): enable `hyper-util/tracing` feature flag
the latest release of `hyper-util` includes hyperium/hyper-util#166,
which changes the behavior of the `TokioExecutor` when the `tracing`
feature flag has been enabled.

when this flag is enabled, the `TokioExecutor` will now propagate the
current tracing span to spawned tasks.

this commit adds this new `hyper-util` feature to the dependency used by
the `kube-client` crate.

* https://github.com/hyperium/hyper-util/pull/166

Signed-off-by: katelyn martin <me+cratelyn@katelyn.world>