feat!: replace chrono with jiff
also drops v1.30 support but we revert this in a follow-up commit before opening the PR (so let's not keep it as part of the commit message as that would be confusing).
v1.30 is dropped as k8s-openapi already dropped it on `main` and we want to test against the `main` status once before reverting to the published version.

Signed-off-by: ngergs <niklas@ngergs.de>