feat!: replace chrono with Jiff (#1868)
* feat!: replace chrono with jiff

also drops v1.30 support but we revert this in a follow-up commit before opening the PR (so let's not keep it as part of the commit message as that would be confusing).
v1.30 is dropped as k8s-openapi already dropped it on `main` and we want to test against the `main` status once before reverting to the published version.

Signed-off-by: ngergs <niklas@ngergs.de>

* fix: remove no longer needed k8s_openapi version check

k8s openapi does not support any version <1.30 so there is no need to check
if we are at a version >=1.30

Signed-off-by: ngergs <niklas@ngergs.de>

* chore: back to published k8s-openapi version

Signed-off-by: ngergs <niklas@ngergs.de>

* fix: revert chrono->jiff replacement that depend on unreleased k8s-openapi

Signed-off-by: ngergs <niklas@ngergs.de>

* fix: unwrap on jiff `round` error that's documented to not happen for our usage

Also aligned all `rfc 3339` comments to call it `RFC3339` so that a string search finds them together with other mentions of this rfc.

Signed-off-by: ngergs <niklas@ngergs.de>

---------

Signed-off-by: ngergs <niklas@ngergs.de>